feat: improve database pylint score
colocate AnnotationModel and ImageModel
